// frontend/src/types/liga.ts
export interface Equipo {
    id: number;
    nombre: string;
    logo_filename?: string;
    color_principal?: string;
    acceso_token?: string;
    liga_id: number;
    puntos_totales: number;
    ganados: number;
    empatados: number;
    perdidos: number;
    puntos_juego_limpio: number;
    puntos_arbitro: number;
    puntos_grada: number;
}

export interface EquipoCreate {
    nombre: string;
    color_principal?: string;
    liga_id: number;
}

export interface EquipoUpdate {
    nombre?: string;
    color_principal?: string;
}

export interface TipoDeporte {
    id: number;
    nombre: string;
    codigo: string;
    tipo_marcador: string;
    permite_empate?: boolean;
    icono?: string;
    // eslint-disable-next-line @typescript-eslint/no-explicit-any
    config?: Record<string, any>;
    descripcion?: string;
    vt_file?: string;
    logo_file?: string;
}
// ...
export interface Marcador {
    goles_local?: number;
    goles_visitante?: number;
    sets_local?: number;
    sets_visitante?: number;
    [key: string]: unknown;
}